Move `display_account_id_maybe` somewhere more sensical
Also deals with some of the pub stuff
